Working dot-dispence. Added commands for setting before/after/suckback-times for the dispencing. Added command for controlling light levels.
git-svn-id: file:///home/srv/www/butt.se/pnp/svn/trunk@95 f672c3ff-8b41-4f22-a4ab-2adcb4f732c7

+++ MachineAPI/MachineController.h
@@ -39,7 +39,7 @@ struct MachineSettings
 	float rotMin;
 	float rotMax;
 	MachineSettings():	xMin(0), xMax(470000), yMin(0), yMax(193000),
-						zMin(0), zMax(10000), rotMin((float)0.0), rotMax((float)2*M_PI) {};
+						zMin(0), zMax(10000), rotMin(0.0f), rotMax(2.0f*M_PI) {};
 };
